odoo_sourcecode_tool/requirements.txt
# Core dependencies
click>=8.1.0
black>=23.0.0
isort>=5.12.0
lxml>=4.9.0
GitPython>=3.1.0
PyYAML>=6.0
rich>=13.0.0
pandas>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
mypy>=1.0.0
ruff>=0.1.0